factorize decode_record and truncate_seq_pair
