+++ extract_features.py
@@ -27,7 +27,7 @@ import modeling
 import tokenization
 import tensorflow as tf
 
-from utils import zero_pad
+from utils import zero_pad, _truncate_seq_pair
 
 flags = tf.flags
 
@@ -281,23 +281,6 @@ def convert_examples_to_features(examples, seq_length, tokenizer):
             input_mask=input_mask,
             input_type_ids=input_type_ids))
   return features
-
-
-def _truncate_seq_pair(tokens_a, tokens_b, max_length):
-  """Truncates a sequence pair in place to the maximum length."""
-
-  # This is a simple heuristic which will always truncate the longer sequence
-  # one token at a time. This makes more sense than truncating an equal percent
-  # of tokens from each, since if one sequence is very short then each token
-  # that's truncated likely contains more information than a longer sequence.
-  while True:
-    total_length = len(tokens_a) + len(tokens_b)
-    if total_length <= max_length:
-      break
-    if len(tokens_a) > len(tokens_b):
-      tokens_a.pop()
-    else:
-      tokens_b.pop()
 
 
 def read_examples(input_file):
