+++ run_pretraining.py
@@ -23,6 +23,8 @@ import modeling
 import optimization
 import tensorflow as tf
 
+from utils import _decode_record
+
 flags = tf.flags
 
 FLAGS = flags.FLAGS
@@ -386,21 +388,6 @@ def input_fn_builder(input_files,
     return d
 
   return input_fn
-
-
-def _decode_record(record, name_to_features):
-  """Decodes a record to a TensorFlow example."""
-  example = tf.parse_single_example(record, name_to_features)
-
-  # tf.Example only supports tf.int64, but the TPU only supports tf.int32.
-  # So cast all int64 to int32.
-  for name in list(example.keys()):
-    t = example[name]
-    if t.dtype == tf.int64:
-      t = tf.to_int32(t)
-    example[name] = t
-
-  return example
 
 
 def main(_):
