+++ run_squad.py
@@ -29,7 +29,7 @@ import tokenization
 import six
 import tensorflow as tf
 
-from utils import zero_pad
+from utils import zero_pad, _decode_record
 
 flags = tf.flags
 
@@ -643,19 +643,6 @@ def input_fn_builder(input_file, seq_length, is_training, drop_remainder):
     name_to_features["start_positions"] = tf.FixedLenFeature([], tf.int64)
     name_to_features["end_positions"] = tf.FixedLenFeature([], tf.int64)
 
-  def _decode_record(record, name_to_features):
-    """Decodes a record to a TensorFlow example."""
-    example = tf.parse_single_example(record, name_to_features)
-
-    # tf.Example only supports tf.int64, but the TPU only supports tf.int32.
-    # So cast all int64 to int32.
-    for name in list(example.keys()):
-      t = example[name]
-      if t.dtype == tf.int64:
-        t = tf.to_int32(t)
-      example[name] = t
-
-    return example
 
   def input_fn(params):
     """The actual input function."""
