+++ utils.py
@@ -1,5 +1,34 @@
+import tensorflow as tf
+
+
 def zero_pad(length, *lists, value=0):
     "Zero-pad the lists up to the specified length"
     for l in lists:
         remaining = length - len(l)
         l += [value] * remaining
+
+
+def _truncate_seq_pair(tokens_a, tokens_b, max_length):
+  """Truncates a sequence pair in place to the maximum length."""
+
+  # This is a simple heuristic which will always truncate the longer sequence
+  # one token at a time. This makes more sense than truncating an equal percent
+  # of tokens from each, since if one sequence is very short then each token
+  # that's truncated likely contains more information than a longer sequence.
+  while len(tokens_a) + len(tokens_b) > max_length:
+    max(tokens_a, tokens_b, key=len).pop()
+
+
+def _decode_record(record, name_to_features):
+    """Decodes a record to a TensorFlow example."""
+    example = tf.parse_single_example(record, name_to_features)
+
+    # tf.Example only supports tf.int64, but the TPU only supports tf.int32.
+    # So cast all int64 to int32.
+    for name in example.keys():
+      t = example[name]
+      if t.dtype == tf.int64:
+        t = tf.to_int32(t)
+      example[name] = t
+
+    return example
